// frontend/src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        ProjectOut: {
            /**
             * Club Id
             * Format: uuid
             */
            club_id: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Description */
            description: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Status */
            status: string;
            /**
             * Tasks
             * @default []
             */
            tasks: components["schemas"]["TaskOut"][];
            /** Title */
            title: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
        TaskOut: {
            /** Active Count */
            active_count?: number | null;
            /** Capacity */
            capacity: number | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Deadline */
            deadline: string | null;
            /** Description */
            description: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Project Id
             * Format: uuid
             */
            project_id: string;
            /** Status */
            status: string;
            /** Title */
            title: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}